Separate out RTL and snapshot expects
