+++ packages/perseus/src/widgets/categorizer/categorizer-validator.test.ts
@@ -1,98 +1,8 @@
-import {screen} from "@testing-library/react";
-import {userEvent as userEventLib} from "@testing-library/user-event";
-
-import {testDependencies} from "../../../../../testing/test-dependencies";
-import * as Dependencies from "../../dependencies";
 import {mockStrings} from "../../strings";
-import {renderQuestion} from "../__testutils__/renderQuestion";
 
-import {Categorizer} from "./categorizer";
-import {question1} from "./categorizer.testdata";
+import categorizerValidator from "./categorizer-validator";
 
 import type {Rubric} from "./categorizer.types";
-import type {APIOptions} from "../../types";
-import type {UserEvent} from "@testing-library/user-event";
-
-describe("categorizer widget", () => {
-    let userEvent: UserEvent;
-    beforeEach(() => {
-        userEvent = userEventLib.setup({
-            advanceTimers: jest.advanceTimersByTime,
-        });
-
-        jest.spyOn(Dependencies, "getDependencies").mockReturnValue(
-            testDependencies,
-        );
-    });
-
-    it("is incorrect when blank", async () => {
-        // Arrange
-        const apiOptions: APIOptions = {
-            isMobile: false,
-        };
-        const {renderer} = renderQuestion(question1, apiOptions);
-
-        // Act
-        const [_, score] = renderer.guessAndScore();
-
-        // Assert
-        expect(score).toMatchInlineSnapshot(`
-            {
-              "message": "Make sure you select something for every row.",
-              "type": "invalid",
-            }
-        `);
-    });
-
-    it("can be answered incorrectly", async () => {
-        // arrange
-        const {renderer} = renderQuestion(question1);
-
-        const firstItem = screen.getAllByRole("row")[0];
-        await userEvent.click(firstItem);
-
-        // act
-        const [_, score] = renderer.guessAndScore();
-
-        // assert
-        expect(score).toMatchInlineSnapshot(`
-            {
-              "message": "Make sure you select something for every row.",
-              "type": "invalid",
-            }
-        `);
-    });
-
-    it("can be answered correctly", async () => {
-        // arrange
-        const {renderer} = renderQuestion(question1);
-
-        // act
-        await userEvent.click(
-            screen.getAllByRole("button", {name: "No relationship"})[0],
-        );
-        await userEvent.click(
-            screen.getAllByRole("button", {
-                name: "Positive linear relationship",
-            })[0],
-        );
-        await userEvent.click(
-            screen.getAllByRole("button", {
-                name: "Negative linear relationship",
-            })[1],
-        );
-        await userEvent.click(
-            screen.getAllByRole("button", {
-                name: "Nonlinear relationship",
-            })[1],
-        );
-
-        renderer.guessAndScore();
-
-        // assert
-        expect(renderer).toHaveBeenAnsweredCorrectly();
-    });
-});
 
 describe("validating answers", () => {
     it("gives points when the answer is correct", () => {
@@ -110,19 +20,17 @@ describe("validating answers", () => {
             static: false,
         };
 
-        const useInput = {
+        const userInput = {
             values: [1, 3],
         } as const;
-        const score = Categorizer.validate(useInput, rubric, mockStrings);
+        const score = categorizerValidator(userInput, rubric, mockStrings);
 
-        expect(score).toMatchInlineSnapshot(`
-            {
-              "earned": 1,
-              "message": null,
-              "total": 1,
-              "type": "points",
-            }
-        `);
+        expect(score).toEqual({
+            earned: 1,
+            message: null,
+            total: 1,
+            type: "points",
+        });
     });
 
     it("does not give points when incorrectly answered", () => {
@@ -140,19 +48,17 @@ describe("validating answers", () => {
             static: false,
         };
 
-        const useInput = {
+        const userInput = {
             values: [2, 3],
         } as const;
-        const score = Categorizer.validate(useInput, rubric, mockStrings);
+        const score = categorizerValidator(userInput, rubric, mockStrings);
 
-        expect(score).toMatchInlineSnapshot(`
-            {
-              "earned": 0,
-              "message": null,
-              "total": 1,
-              "type": "points",
-            }
-        `);
+        expect(score).toEqual({
+            earned: 0,
+            message: null,
+            total: 1,
+            type: "points",
+        });
     });
 
     it("tells the learner its not complete if not selected", () => {
@@ -170,16 +76,14 @@ describe("validating answers", () => {
             static: false,
         };
 
-        const useInput = {
+        const userInput = {
             values: [2],
         } as const;
-        const score = Categorizer.validate(useInput, rubric, mockStrings);
+        const score = categorizerValidator(userInput, rubric, mockStrings);
 
-        expect(score).toMatchInlineSnapshot(`
-            {
-              "message": "Make sure you select something for every row.",
-              "type": "invalid",
-            }
-        `);
+        expect(score).toEqual({
+            message: "Make sure you select something for every row.",
+            type: "invalid",
+        });
     });
 });
