+++ packages/perseus/src/widgets/categorizer/categorizer.test.ts
@@ -1,3 +1,6 @@
+import {screen} from "@testing-library/react";
+import {userEvent as userEventLib} from "@testing-library/user-event";
+
 import {testDependencies} from "../../../../../testing/test-dependencies";
 import * as Dependencies from "../../dependencies";
 import {renderQuestion} from "../__testutils__/renderQuestion";
@@ -6,9 +9,15 @@ import {Categorizer} from "./categorizer";
 import {question1} from "./categorizer.testdata";
 
 import type {APIOptions} from "../../types";
+import type {UserEvent} from "@testing-library/user-event";
 
 describe("categorizer widget", () => {
+    let userEvent: UserEvent;
     beforeEach(() => {
+        userEvent = userEventLib.setup({
+            advanceTimers: jest.advanceTimersByTime,
+        });
+
         jest.spyOn(Dependencies, "getDependencies").mockReturnValue(
             testDependencies,
         );
@@ -38,6 +47,74 @@ describe("categorizer widget", () => {
 
         // Assert
         expect(container).toMatchSnapshot("first mobile render");
+    });
+
+    it("is incorrect when blank", async () => {
+        // Arrange
+        const apiOptions: APIOptions = {
+            isMobile: false,
+        };
+        const {renderer} = renderQuestion(question1, apiOptions);
+
+        // Act
+        const [_, score] = renderer.guessAndScore();
+
+        // Assert
+        expect(score).toMatchInlineSnapshot(`
+            {
+              "message": "Make sure you select something for every row.",
+              "type": "invalid",
+            }
+        `);
+    });
+
+    it("can be answered incorrectly", async () => {
+        // arrange
+        const {renderer} = renderQuestion(question1);
+
+        const firstItem = screen.getAllByRole("row")[0];
+        await userEvent.click(firstItem);
+
+        // act
+        const [_, score] = renderer.guessAndScore();
+
+        // assert
+        expect(score).toMatchInlineSnapshot(`
+            {
+              "message": "Make sure you select something for every row.",
+              "type": "invalid",
+            }
+        `);
+    });
+
+    it("can be answered correctly", async () => {
+        // arrange
+        const {renderer} = renderQuestion(question1);
+
+        // act
+        await userEvent.click(
+            screen.getAllByRole("button", {name: "No relationship"})[0],
+        );
+        await userEvent.click(
+            screen.getAllByRole("button", {
+                name: "Positive linear relationship",
+            })[0],
+        );
+        await userEvent.click(
+            screen.getAllByRole("button", {
+                name: "Negative linear relationship",
+            })[1],
+        );
+        await userEvent.click(
+            screen.getAllByRole("button", {
+                name: "Nonlinear relationship",
+            })[1],
+        );
+
+        renderer.guessAndScore();
+
+        // assert
+        expect(renderer).toHaveBeenAnsweredCorrectly();
     });
 
     it("can get user input from props", () => {
